comment the table

+++ utils/dice_score.py
@@ -29,6 +29,6 @@ def dice_loss(input: Tensor, target: Tensor, multiclass: bool = False):
 
 def UnetLoss(preds, targets):
     ce_loss = nn.CrossEntropyLoss()(preds, targets)
-    print((torch.max(preds, 1)[1] == targets).astype(int))
+    print((torch.max(preds, 1)[1] == targets).int())
     acc = (torch.max(preds, 1)[1] == targets).float().mean()
     return ce_loss, acc
